+++ src/routes.py
@@ -5,11 +5,12 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from flask import Blueprint
-from flask import make_response, redirect, render_template, request, url_for
+from flask import redirect, render_template, request, url_for
 
 from src.database import session_var
-from src.forms import LoginForm, PostForm, RegistrationForm, TopicForm
-from src.models import Topic, User, UserSession
+from src.forms import PostForm, TopicForm
+from src.models import Topic
+from src.users.models import User, UserSession
 
 if TYPE_CHECKING:
     from werkzeug.wrappers.response import Response
@@ -48,54 +49,20 @@ def index() -> str:
     return render_template("index.html", users=users, current_user=current_user)
 
 
-@bp.route("/registration", methods=["POST", "GET"])
-def registration() -> str | Response:
-    """Handle user's registration form."""
-    form = RegistrationForm()
-    if form.validate_on_submit():
-        User.create_user(form.username.data, form.password.data)
-        return redirect(url_for("routes.login"))
-    return render_template("registration.html", form=form)
-
-
-@bp.route("/login", methods=["POST", "GET"])
-def login() -> str | Response:
-    """Handle user's login form."""
-    form = LoginForm()
-    if form.validate_on_submit() and (user := User.get_user_by_credentials(form.username.data, form.password.data)):
-        user_session = user.create_session()
-        response = make_response(redirect(url_for("routes.topics")))
-        response.set_cookie("session_id", user_session.session_id)
-        return response
-    return render_template("login.html", form=form)
-
-
-@bp.route("/logout")
-def logout() -> Response:
-    """Log out users."""
-    session_id = request.cookies.get("session_id")
-    session_to_delete = None
-    if session_id is not None:
-        session_to_delete = UserSession.get_user_session_by_session_id(session_id)
-    if session_to_delete:
-        session_to_delete.delete()
-    return redirect(url_for("routes.login"))
-
-
 @bp.route("/topics")
 def topics() -> str | Response:
     """Handle topics page."""
     if (current_user := get_current_user()) is not None:
         topic_list = Topic.get_topics()
         return render_template("topics.html", topic_list=topic_list, current_user=current_user)
-    return redirect(url_for("routes.login"))
+    return redirect(url_for("users.login"))
 
 
 @bp.route("/topics/create", methods=["POST", "GET"])
 def create_topic() -> str | Response:
     """Handle create topic page."""
     if (current_user := get_current_user()) is None:
-        return redirect(url_for("routes.login"))
+        return redirect(url_for("users.login"))
     form = TopicForm()
     if not form.validate_on_submit():
         return render_template("topics-create.html", form=form, current_user=current_user)
@@ -107,7 +74,7 @@ def create_topic() -> str | Response:
 def topic_page(topic_id: int) -> str | Response:
     """Handle topic page."""
     if (current_user := get_current_user()) is None:
-        return redirect(url_for("routes.login"))
+        return redirect(url_for("users.login"))
     if not (topic := Topic.get(topic_id)):
         return render_template("404.html", current_user=current_user)
     return render_template("topic.html", current_user=current_user, topic=topic)
@@ -117,7 +84,7 @@ def topic_page(topic_id: int) -> str | Response:
 def create_post(topic_id: int) -> str | Response:  # noqa: CFQ004
     """Handle post creation page."""
     if (current_user := get_current_user()) is None:
-        return redirect(url_for("routes.login"))
+        return redirect(url_for("users.login"))
     form = PostForm()
     if not (topic := Topic.get(topic_id)):
         return render_template("404.html", current_user=current_user)
