+++ src/users/__init__.py
@@ -1,0 +1,3 @@
+"""'users' package's initialization module."""
+
+from . import routes  # noqa
